+++ java/src/main/java/com/github/cowwoc/requirements/java/internal/extension/AbstractObjectValidator.java
@@ -109,12 +109,6 @@ public abstract class AbstractObjectValidator<S, T> implements ExtensibleObjectV
 	}
 
 	@Override
-	public boolean isActualAvailable()
-	{
-		return true;
-	}
-
-	@Override
 	public T getActual()
 	{
 		return actual;